refactor: create true controller
