+++ src/components/opt-in-out/helpers/index.ts
@@ -1,4 +1,3 @@
 export * from './post-form-data'
-export * from './get-io-redirect-path'
-export * from './get-learn-redirect-path'
 export * from './make-beta-welcome-toast'
+export * from './platform-options'
