+++ src/components/opt-in-out/index.tsx
@@ -1,60 +1,21 @@
-import { useCallback, useEffect, useState } from 'react'
+import { useCallback, useEffect } from 'react'
 import { useRouter } from 'next/router'
 import Cookies from 'js-cookie'
 import { v4 as uuid } from 'uuid'
-import { IconSignOut16 } from '@hashicorp/flight-icons/svg-react/sign-out-16'
 import '@reach/dialog/styles.css'
 import { safeAnalyticsTrack } from 'lib/analytics'
-import Button from 'components/button'
-import Dialog from 'components/dialog'
-import OptOutForm from './components/opt-out-form'
-import {
-  PlatformOptionRedirectData,
-  OptInOutProps,
-  isOptInPlatformOption,
-} from './types'
-import {
-  getIoRedirectPath,
-  getLearnRedirectPath,
-  postFormData,
-  makeBetaWelcomeToast,
-} from './helpers'
+import { OptInOutProps, isOptInPlatformOption } from './types'
+import { PLATFORM_OPTIONS, postFormData, makeBetaWelcomeToast } from './helpers'
 import { OptOutFormState } from './components/opt-out-form/types'
+import { OptOutButtonAndDialog } from './components/opt-out-button-and-dialog'
 
-// Could these go in the config? or I could source the base urls elsewhere
-export const PLATFORM_OPTIONS: PlatformOptionRedirectData = {
-  learn: {
-    base_url: 'https://learn-git-ksspike-opt-in-redirects-hashicorp.vercel.app', // FOR TESTING PURPOSES NEED TO UPDATE for - 'https://learn.hashicorp.com/'
-    getRedirectPath: getLearnRedirectPath,
-    cookieKey: 'learn-beta-opt-in',
-    cookieAnalyticsKey: 'learn-beta-opt-in-tracked',
-  },
-  'waypoint-io': {
-    base_url: 'https://www.waypointproject.io/',
-    getRedirectPath(path) {
-      return this.base_url + getIoRedirectPath(path)
-    },
-    cookieKey: 'waypoint-io-beta-opt-in',
-    cookieAnalyticsKey: 'waypoint-io-beta-opt-in-tracked',
-  },
-  'vault-io': {
-    base_url: 'https://www.vaultproject.io/',
-    getRedirectPath(path) {
-      return this.base_url + getIoRedirectPath(path)
-    },
-    cookieKey: 'vault-io-beta-opt-in',
-    cookieAnalyticsKey: 'vault-io-beta-opt-in-tracked',
-  },
-}
-
-export default function OptInOut({ platform, redirectPath }: OptInOutProps) {
+export default function OptInOutController({
+  platform,
+  redirectPath,
+}: OptInOutProps) {
   // fire toast, render button, etc
   const router = useRouter()
-  const [showDialog, setShowDialog] = useState(false)
   const optedIn = Cookies.get(PLATFORM_OPTIONS[platform].cookieKey)
-
-  const openDialog = () => setShowDialog(true)
-  const closeDialog = () => setShowDialog(false)
   const url =
     redirectPath || PLATFORM_OPTIONS[platform].getRedirectPath(router.asPath)
 
@@ -68,16 +29,19 @@ export default function OptInOut({ platform, redirectPath }: OptInOutProps) {
         segment_anonymous_id: uuid(),
         primary_opt_out_reason: state.optOutReason,
         details: state.optOutDetails,
-        opt_out_page_url: router.asPath,
+        opt_out_page_url: new URL(
+          router.asPath,
+          __config.dev_dot.canonical_base_url
+        ).toString(),
       })
-      // Cookies.remove(PLATFORM_OPTIONS[platform].cookieKey)
-      // Cookies.remove(PLATFORM_OPTIONS[platform].cookieAnalyticsKey)
       // safeAnalyticsTrack('Beta Opted Out', {
       //   bucket: platform,
       // })
+      // Cookies.remove(PLATFORM_OPTIONS[platform].cookieKey)
+      // Cookies.remove(PLATFORM_OPTIONS[platform].cookieAnalyticsKey)
       // window.location.assign(url)
     },
-    [platform, url]
+    [platform, url, router.asPath]
   )
 
   /**
@@ -96,22 +60,5 @@ export default function OptInOut({ platform, redirectPath }: OptInOutProps) {
   //   return null
   // }
 
-  return (
-    <div>
-      <Button
-        color="tertiary"
-        text="Leave Beta"
-        icon={<IconSignOut16 />}
-        iconPosition="trailing"
-        onClick={openDialog}
-      />
-      <Dialog onDismiss={closeDialog} isOpen={showDialog} label="Opt out form">
-        <OptOutForm
-          onSubmit={handleOptOut}
-          onDismiss={closeDialog}
-          platform="learn"
-        />
-      </Dialog>
-    </div>
-  )
+  return <OptOutButtonAndDialog handleOptOut={handleOptOut} />
 }
